make entire bitfield unsupported
