+++ packages/flutter/lib/src/foundation/basic_types.dart
@@ -4,13 +4,11 @@
 
 import 'dart:async';
 import 'dart:collection';
-import 'internal.dart'
-  if (dart.html) 'internal_web.dart';
 
 // COMMON SIGNATURES
 
 export 'dart:ui' show VoidCallback;
-export 'internal.dart'
+export 'bitfield.dart'
   if (dart.html) 'internal_web.dart';
 
 /// Signature for callbacks that report that an underlying value has changed.
@@ -69,64 +67,6 @@ typedef AsyncValueSetter<T> = Future<void> Function(T value);
 ///  * [ValueGetter], a synchronous version of this signature.
 ///  * [AsyncValueSetter], the setter equivalent of this signature.
 typedef AsyncValueGetter<T> = Future<T> Function();
-
-
-// BITFIELD
-
-/// A BitField over an enum (or other class whose values implement "index").
-/// Only the first 62 values of the enum can be used as indices.
-class BitField<T extends dynamic> {
-  /// Creates a bit field of all zeros.
-  ///
-  /// The given length must be at most 62.
-  BitField(this._length)
-    : assert(_length <= _smiBits),
-      _bits = _allZeros;
-
-  /// Creates a bit field filled with a particular value.
-  ///
-  /// If the value argument is true, the bits are filled with ones. Otherwise,
-  /// the bits are filled with zeros.
-  ///
-  /// The given length must be at most 62.
-  BitField.filled(this._length, bool value)
-    : assert(_length <= _smiBits),
-      _bits = value ? _allOnes : _allZeros;
-
-  final int _length;
-  int _bits;
-
-  static const int _smiBits = 62; // see https://www.dartlang.org/articles/numeric-computation/#smis-and-mints
-  static const int _allZeros = 0;
-  static const int _allOnes = kMaxUnsignedSMI; // 2^(_kSMIBits+1)-1
-
-  /// Returns whether the bit with the given index is set to one.
-  bool operator [](T index) {
-    assert(index.index < _length);
-    return (_bits & 1 << index.index) > 0;
-  }
-
-  /// Sets the bit with the given index to the given value.
-  ///
-  /// If value is true, the bit with the given index is set to one. Otherwise,
-  /// the bit is set to zero.
-  void operator []=(T index, bool value) {
-    assert(index.index < _length);
-    if (value)
-      _bits = _bits | (1 << index.index);
-    else
-      _bits = _bits & ~(1 << index.index);
-  }
-
-  /// Sets all the bits to the given value.
-  ///
-  /// If the value is true, the bits are all set to one. Otherwise, the bits are
-  /// all set to zero. Defaults to setting all the bits to zero.
-  void reset([ bool value = false ]) {
-    _bits = value ? _allOnes : _allZeros;
-  }
-}
-
 
 // LAZY CACHING ITERATOR
